PUB-1765 - Fixed circular dependencies
